Prefer the `For #..` syntax instead of `Closes #..` in naming patches
There are projects for which the `Closes #..` syntax may break the process.
For example on Fenix tickets are to be closed by QA only after verifying the
patches.
`Closes #..` should only be used when we know for a fact that we want the issue
be closed automatically after the PR merge.

For external contributors, new to Mozilla projects the default patches naming
should use the `For #..` syntax.
It's easy to manually close the ticket after the PR merge if needed and so
prevent adding to the ticket history unnecessarily closes and reopenings when
the issue should've not be automatically closed.